+++ addons/es/sii/tax_combo_test.go
@@ -3,6 +3,7 @@ package sii
 import (
 	"testing"
 
+	"github.com/invopop/gobl/cbc"
 	"github.com/invopop/gobl/l10n"
 	"github.com/invopop/gobl/num"
 	"github.com/invopop/gobl/regimes/es"
@@ -90,7 +91,7 @@ func TestNormalizeTaxCombo(t *testing.T) {
 		tc := &tax.Combo{
 			Category: tax.CategoryVAT,
 			Key:      tax.KeyOutsideScope,
-			Ext: tax.ExtensionsOf(tax.ExtMap{
+			Ext: tax.ExtensionsOf(cbc.CodeMap{
 				ExtKeyOutsideScope: "other",
 			}),
 		}
@@ -125,7 +126,7 @@ func TestNormalizeTaxCombo(t *testing.T) {
 		tc := &tax.Combo{
 			Category: tax.CategoryVAT,
 			Rate:     tax.RateGeneral,
-			Ext: tax.ExtensionsOf(tax.ExtMap{
+			Ext: tax.ExtensionsOf(cbc.CodeMap{
 				ExtKeyRegime: "03",
 			}),
 		}
@@ -136,7 +137,7 @@ func TestNormalizeTaxCombo(t *testing.T) {
 	t.Run("with exempt code set", func(t *testing.T) {
 		tc := &tax.Combo{
 			Category: tax.CategoryVAT,
-			Ext: tax.ExtensionsOf(tax.ExtMap{
+			Ext: tax.ExtensionsOf(cbc.CodeMap{
 				ExtKeyExempt: "E6",
 			}),
 		}
@@ -148,7 +149,7 @@ func TestNormalizeTaxCombo(t *testing.T) {
 	t.Run("with export code set", func(t *testing.T) {
 		tc := &tax.Combo{
 			Category: tax.CategoryVAT,
-			Ext: tax.ExtensionsOf(tax.ExtMap{
+			Ext: tax.ExtensionsOf(cbc.CodeMap{
 				ExtKeyExempt: "E2",
 				ExtKeyRegime: "02",
 			}),
@@ -169,7 +170,7 @@ func TestNormalizeTaxCombo(t *testing.T) {
 	t.Run("with outside-scope", func(t *testing.T) {
 		tc := &tax.Combo{
 			Category: tax.CategoryVAT,
-			Ext: tax.ExtensionsOf(tax.ExtMap{
+			Ext: tax.ExtensionsOf(cbc.CodeMap{
 				ExtKeyOutsideScope: "location",
 				ExtKeyRegime:       "01",
 			}),
@@ -182,7 +183,7 @@ func TestNormalizeTaxCombo(t *testing.T) {
 	t.Run("with intra-community", func(t *testing.T) {
 		tc := &tax.Combo{
 			Category: tax.CategoryVAT,
-			Ext: tax.ExtensionsOf(tax.ExtMap{
+			Ext: tax.ExtensionsOf(cbc.CodeMap{
 				ExtKeyExempt: "E5",
 				ExtKeyRegime: "01",
 			}),
@@ -202,7 +203,7 @@ func TestValidateTaxCombo(t *testing.T) {
 			Category: tax.CategoryVAT,
 			Rate:     tax.RateGeneral,
 			Percent:  num.NewPercentage(210, 3),
-			Ext: tax.ExtensionsOf(tax.ExtMap{
+			Ext: tax.ExtensionsOf(cbc.CodeMap{
 				ExtKeyRegime: "01",
 			}),
 		}
@@ -223,7 +224,7 @@ func TestValidateTaxCombo(t *testing.T) {
 		tc := &tax.Combo{
 			Category: tax.CategoryVAT,
 			Key:      tax.KeyExempt,
-			Ext: tax.ExtensionsOf(tax.ExtMap{
+			Ext: tax.ExtensionsOf(cbc.CodeMap{
 				ExtKeyRegime: "01",
 				ExtKeyExempt: "E1",
 			}),
@@ -235,7 +236,7 @@ func TestValidateTaxCombo(t *testing.T) {
 	t.Run("excludes E2 exemption code with regime 01", func(t *testing.T) {
 		tc := &tax.Combo{
 			Category: tax.CategoryVAT,
-			Ext: tax.ExtensionsOf(tax.ExtMap{
+			Ext: tax.ExtensionsOf(cbc.CodeMap{
 				ExtKeyRegime: "01",
 				ExtKeyExempt: "E2",
 			}),
@@ -247,7 +248,7 @@ func TestValidateTaxCombo(t *testing.T) {
 	t.Run("excludes E3 exemption code with regime 01", func(t *testing.T) {
 		tc := &tax.Combo{
 			Category: tax.CategoryVAT,
-			Ext: tax.ExtensionsOf(tax.ExtMap{
+			Ext: tax.ExtensionsOf(cbc.CodeMap{
 				ExtKeyRegime: "01",
 				ExtKeyExempt: "E3",
 			}),
@@ -260,7 +261,7 @@ func TestValidateTaxCombo(t *testing.T) {
 	t.Run("allows E2 exemption code with non-01 regime", func(t *testing.T) {
 		tc := &tax.Combo{
 			Category: tax.CategoryVAT,
-			Ext: tax.ExtensionsOf(tax.ExtMap{
+			Ext: tax.ExtensionsOf(cbc.CodeMap{
 				ExtKeyRegime: "02",
 				ExtKeyExempt: "E2",
 			}),
@@ -272,7 +273,7 @@ func TestValidateTaxCombo(t *testing.T) {
 	t.Run("excludes E2 exemption code with regime 01 and IGIC category", func(t *testing.T) {
 		tc := &tax.Combo{
 			Category: es.TaxCategoryIGIC,
-			Ext: tax.ExtensionsOf(tax.ExtMap{
+			Ext: tax.ExtensionsOf(cbc.CodeMap{
 				ExtKeyRegime: "01",
 				ExtKeyExempt: "E2",
 			}),
@@ -287,7 +288,7 @@ func TestValidateTaxCombo(t *testing.T) {
 			Category: tax.CategoryVAT,
 			Rate:     tax.RateGeneral,
 			Percent:  num.NewPercentage(210, 3),
-			Ext: tax.ExtensionsOf(tax.ExtMap{
+			Ext: tax.ExtensionsOf(cbc.CodeMap{
 				ExtKeyProduct: "goods",
 			}),
 		}
@@ -301,7 +302,7 @@ func TestValidateTaxCombo(t *testing.T) {
 			Category: tax.CategoryVAT,
 			Rate:     tax.RateGeneral,
 			Percent:  num.NewPercentage(210, 3),
-			Ext: tax.ExtensionsOf(tax.ExtMap{
+			Ext: tax.ExtensionsOf(cbc.CodeMap{
 				ExtKeyRegime: "01",
 				ExtKeyExempt: "E1",
 			}),
@@ -314,7 +315,7 @@ func TestValidateTaxCombo(t *testing.T) {
 	t.Run("allows only one of outside scope or exempt", func(t *testing.T) {
 		tc := &tax.Combo{
 			Category: tax.CategoryVAT,
-			Ext: tax.ExtensionsOf(tax.ExtMap{
+			Ext: tax.ExtensionsOf(cbc.CodeMap{
 				ExtKeyRegime:       "01",
 				ExtKeyOutsideScope: "location",
 				ExtKeyExempt:       "E1",
@@ -328,7 +329,7 @@ func TestValidateTaxCombo(t *testing.T) {
 	t.Run("valid not subject when percent is nil", func(t *testing.T) {
 		tc := &tax.Combo{
 			Category: tax.CategoryVAT,
-			Ext: tax.ExtensionsOf(tax.ExtMap{
+			Ext: tax.ExtensionsOf(cbc.CodeMap{
 				ExtKeyRegime:       "01",
 				ExtKeyOutsideScope: "location",
 			}),
@@ -342,7 +343,7 @@ func TestValidateTaxCombo(t *testing.T) {
 			Category: tax.CategoryVAT,
 			Rate:     tax.RateGeneral,
 			Percent:  num.NewPercentage(210, 3),
-			Ext: tax.ExtensionsOf(tax.ExtMap{
+			Ext: tax.ExtensionsOf(cbc.CodeMap{
 				ExtKeyRegime:       "01",
 				ExtKeyOutsideScope: "location",
 			}),
@@ -356,7 +357,7 @@ func TestValidateTaxCombo(t *testing.T) {
 			Category: tax.CategoryVAT,
 			Rate:     tax.RateGeneral,
 			Percent:  num.NewPercentage(210, 3),
-			Ext: tax.ExtensionsOf(tax.ExtMap{
+			Ext: tax.ExtensionsOf(cbc.CodeMap{
 				ExtKeyRegime: "01",
 			}),
 		}
@@ -367,7 +368,7 @@ func TestValidateTaxCombo(t *testing.T) {
 	t.Run("valid exempt with no percent", func(t *testing.T) {
 		tc := &tax.Combo{
 			Category: tax.CategoryVAT,
-			Ext: tax.ExtensionsOf(tax.ExtMap{
+			Ext: tax.ExtensionsOf(cbc.CodeMap{
 				ExtKeyRegime: "01",
 				ExtKeyExempt: "E1",
 			}),