+++ tax/totals_calculator_test.go
@@ -278,7 +278,7 @@ func TestTotalBySumCalculate(t *testing.T) {
 						{
 							Category: tax.CategoryVAT,
 							Key:      tax.KeyExempt,
-							Ext: tax.ExtensionsOf(tax.ExtMap{
+							Ext: tax.ExtensionsOf(cbc.CodeMap{
 								tbai.ExtKeyExempt: "E1",
 							}),
 						},
@@ -295,7 +295,7 @@ func TestTotalBySumCalculate(t *testing.T) {
 						Rates: []*tax.RateTotal{
 							{
 								Key: tax.KeyExempt,
-								Ext: tax.ExtensionsOf(tax.ExtMap{
+								Ext: tax.ExtensionsOf(cbc.CodeMap{
 									tbai.ExtKeyExempt: "E1",
 								}),
 								Base:    num.MakeAmount(10000, 2),
@@ -409,7 +409,7 @@ func TestTotalBySumCalculate(t *testing.T) {
 							Category: tax.CategoryVAT,
 							Key:      tax.KeyStandard,
 							Rate:     tax.RateGeneral,
-							Ext: tax.ExtensionsOf(tax.ExtMap{
+							Ext: tax.ExtensionsOf(cbc.CodeMap{
 								pt.ExtKeyRegion:    "PT-AC",
 								saft.ExtKeyTaxRate: "NOR",
 							}),
@@ -430,7 +430,7 @@ func TestTotalBySumCalculate(t *testing.T) {
 								Base:    num.MakeAmount(10000, 2),
 								Percent: num.NewPercentage(160, 3),
 								Amount:  num.MakeAmount(1600, 2),
-								Ext: tax.ExtensionsOf(tax.ExtMap{
+								Ext: tax.ExtensionsOf(cbc.CodeMap{
 									pt.ExtKeyRegion:    "PT-AC",
 									saft.ExtKeyTaxRate: "NOR",
 								}),
@@ -1069,7 +1069,7 @@ func TestTotalBySumCalculate(t *testing.T) {
 						{
 							Category: tax.CategoryVAT,
 							Key:      tax.KeyExempt,
-							Ext: tax.ExtensionsOf(tax.ExtMap{
+							Ext: tax.ExtensionsOf(cbc.CodeMap{
 								tbai.ExtKeyExempt: "E1",
 							}),
 						},
@@ -1085,7 +1085,7 @@ func TestTotalBySumCalculate(t *testing.T) {
 						Rates: []*tax.RateTotal{
 							{
 								Key: tax.KeyExempt,
-								Ext: tax.ExtensionsOf(tax.ExtMap{
+								Ext: tax.ExtensionsOf(cbc.CodeMap{
 									tbai.ExtKeyExempt: "E1",
 								}),
 								Base:   num.MakeAmount(10000, 2),
@@ -1105,7 +1105,7 @@ func TestTotalBySumCalculate(t *testing.T) {
 					taxes: tax.Set{
 						{
 							Category: tax.CategoryVAT,
-							Ext: tax.ExtensionsOf(tax.ExtMap{
+							Ext: tax.ExtensionsOf(cbc.CodeMap{
 								tbai.ExtKeyExempt: "E1",
 							}),
 						},
@@ -1116,7 +1116,7 @@ func TestTotalBySumCalculate(t *testing.T) {
 					taxes: tax.Set{
 						{
 							Category: tax.CategoryVAT,
-							Ext: tax.ExtensionsOf(tax.ExtMap{
+							Ext: tax.ExtensionsOf(cbc.CodeMap{
 								tbai.ExtKeyExempt: "E1",
 							}),
 						},
@@ -1132,7 +1132,7 @@ func TestTotalBySumCalculate(t *testing.T) {
 						Rates: []*tax.RateTotal{
 							{
 								Key: tax.KeyStandard,
-								Ext: tax.ExtensionsOf(tax.ExtMap{
+								Ext: tax.ExtensionsOf(cbc.CodeMap{
 									tbai.ExtKeyExempt: "E1",
 								}),
 								Base:   num.MakeAmount(12000, 2),
@@ -1162,7 +1162,7 @@ func TestTotalBySumCalculate(t *testing.T) {
 						{
 							Category: tax.CategoryVAT,
 							Key:      tax.KeyExempt,
-							Ext: tax.ExtensionsOf(tax.ExtMap{
+							Ext: tax.ExtensionsOf(cbc.CodeMap{
 								tbai.ExtKeyExempt: "E2",
 							}),
 						},
@@ -1184,7 +1184,7 @@ func TestTotalBySumCalculate(t *testing.T) {
 							},
 							{
 								Key: tax.KeyExempt,
-								Ext: tax.ExtensionsOf(tax.ExtMap{
+								Ext: tax.ExtensionsOf(cbc.CodeMap{
 									tbai.ExtKeyExempt: "E2",
 								}),
 								Base:   num.MakeAmount(10000, 2),
@@ -1210,7 +1210,7 @@ func TestTotalBySumCalculate(t *testing.T) {
 						},
 						{
 							Category: it.TaxCategoryIRPEF,
-							Ext: tax.ExtensionsOf(tax.ExtMap{
+							Ext: tax.ExtensionsOf(cbc.CodeMap{
 								sdi.ExtKeyRetained: "A",
 							}),
 							Percent: num.NewPercentage(20, 2),
@@ -1226,7 +1226,7 @@ func TestTotalBySumCalculate(t *testing.T) {
 						},
 						{
 							Category: it.TaxCategoryIRPEF,
-							Ext: tax.ExtensionsOf(tax.ExtMap{
+							Ext: tax.ExtensionsOf(cbc.CodeMap{
 								sdi.ExtKeyRetained: "J", // truffles!
 							}),
 							Percent: num.NewPercentage(20, 2),
@@ -1254,7 +1254,7 @@ func TestTotalBySumCalculate(t *testing.T) {
 						Retained: true,
 						Rates: []*tax.RateTotal{
 							{
-								Ext: tax.ExtensionsOf(tax.ExtMap{
+								Ext: tax.ExtensionsOf(cbc.CodeMap{
 									sdi.ExtKeyRetained: "A",
 								}),
 								Base:    num.MakeAmount(10000, 2),
@@ -1262,7 +1262,7 @@ func TestTotalBySumCalculate(t *testing.T) {
 								Amount:  num.MakeAmount(2000, 2),
 							},
 							{
-								Ext: tax.ExtensionsOf(tax.ExtMap{
+								Ext: tax.ExtensionsOf(cbc.CodeMap{
 									sdi.ExtKeyRetained: "J",
 								}),
 								Base:    num.MakeAmount(10000, 2),